refactored app model to support async/await
- broke old `Model.swift` into various subcomponents, many of which are actors
- update the UI code to support the model updates
- some minor local state model changes along the way
- some WIP commented code for now
