+++ InventoryWatch/Helpers.swift
@@ -7,6 +7,9 @@
 
 import Foundation
 
+typealias SKUString = String
+typealias CountryCode = String
+
 extension Array where Element == String {
   func sortedNumerically() -> [Element] {
     sorted { lhs, rhs in
@@ -39,11 +42,8 @@ public extension URLSession {
     func data(for request: URLRequest) async throws -> (Data, URLResponse) {
         var dataTask: URLSessionDataTask?
         let onCancel = { dataTask?.cancel() }
-
+        
         return try await withTaskCancellationHandler(
-            handler: {
-                onCancel()
-            },
             operation: {
                 try await withCheckedThrowingContinuation { continuation in
                     dataTask = self.dataTask(with: request) { data, response, error in
@@ -51,13 +51,14 @@ public extension URLSession {
                             let error = error ?? URLError(.badServerResponse)
                             return continuation.resume(throwing: error)
                         }
-
+                        
                         continuation.resume(returning: (data, response))
                     }
-
+                    
                     dataTask?.resume()
                 }
-            }
+            },
+            onCancel: { [dataTask] in dataTask?.cancel() }
         )
     }
 }
