+++ InventoryWatch/Model/PartAvailability.swift
@@ -13,6 +13,7 @@ struct PartAvailability: Equatable, Hashable {
     }
     
     let partNumber: String
+    let partName: String
     let availability: PickupAvailability
 }
